+++ src/main/scala/au/com/agl/arc/util/SQLUtils.scala
@@ -27,15 +27,22 @@ object SQLUtils {
     // using regex from the apache zeppelin project
     val stmt = sqlParams.foldLeft(sql) {
       case (sql, (k,v)) => { 
-          val regex = "[$][{]\\s*" + k + "\\s*(?:=[^}]+)?[}]"
+          val placeholderRegex = "[$][{]\\s*" + k + "\\s*(?:=[^}]+)?[}]"
 
           // throw error if no match found
-          if (regex.r.findAllIn(sql).length == 0) {
-            throw new Exception(s"No placeholder found in SQL statement for sqlParam: '${k}'")
+          if (placeholderRegex.r.findAllIn(sql).length == 0) {
+            throw new Exception(s"No placeholder found in SQL statement for sqlParam: '${k}'.")
           }           
 
-          regex.r.replaceAllIn(sql, v)
+          placeholderRegex.r.replaceAllIn(sql, v)
       }
+    }
+
+    // find missing replacements
+    val missingRegex = "[$][{](\\w*)(?:=[^}]+)?[}]"
+    val missingValues = missingRegex.r.findAllIn(stmt).toList
+    if (missingValues.length != 0) {
+      throw new Exception(s"""No replacement value found in sqlParams ${sqlParams.keys.toList.mkString("[", ", ", "]")} for placeholders: ${missingValues.mkString("[", ", ", "]")}.""")
     }
 
     logger.trace()