+++ src/test/scala/au/com/agl/arc/extract/XMLExtractSuite.scala
@@ -80,7 +80,6 @@ class XMLExtractSuite extends FunSuite with BeforeAndAfter {
         persist=false,
         numPartitions=None,
         partitionBy=Nil,
-        basePath=None,
         contiguousIndex=true
       )
     ).get
@@ -120,7 +119,6 @@ class XMLExtractSuite extends FunSuite with BeforeAndAfter {
         persist=false,
         numPartitions=None,
         partitionBy=Nil,
-        basePath=None,
         contiguousIndex=true
       )
     )
@@ -139,7 +137,6 @@ class XMLExtractSuite extends FunSuite with BeforeAndAfter {
         persist=true,
         numPartitions=None,
         partitionBy=Nil,
-        basePath=None,
         contiguousIndex=true
       )
     )
@@ -179,7 +176,6 @@ class XMLExtractSuite extends FunSuite with BeforeAndAfter {
           persist=false,
           numPartitions=None,
           partitionBy=Nil,
-          basePath=None,
           contiguousIndex=true
         )
       )
@@ -200,7 +196,6 @@ class XMLExtractSuite extends FunSuite with BeforeAndAfter {
           persist=false,
           numPartitions=None,
           partitionBy=Nil,
-          basePath=None,
           contiguousIndex=true
         )
       )
@@ -220,7 +215,6 @@ class XMLExtractSuite extends FunSuite with BeforeAndAfter {
         persist=false,
         numPartitions=None,
         partitionBy=Nil,
-        basePath=None,
         contiguousIndex=true
       )
     ).get
@@ -250,7 +244,6 @@ class XMLExtractSuite extends FunSuite with BeforeAndAfter {
         persist=false,
         numPartitions=None,
         partitionBy=Nil,
-        basePath=None,
         contiguousIndex=true
       )
     ).get
@@ -277,7 +270,6 @@ class XMLExtractSuite extends FunSuite with BeforeAndAfter {
         persist=false,
         numPartitions=None,
         partitionBy=Nil,
-        basePath=None,
         contiguousIndex=true
       )
     ).get
@@ -312,7 +304,6 @@ class XMLExtractSuite extends FunSuite with BeforeAndAfter {
         persist=false,
         numPartitions=None,
         partitionBy=Nil,
-        basePath=None,
         contiguousIndex=true
       )
     ).get